+++ backend/src/main/java/org/donggle/backend/application/service/connection/notion/dto/NotionOwnerResponse.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.application.service.oauth.notion.dto;
+package org.donggle.backend.application.service.connection.notion.dto;
 
 public record NotionOwnerResponse(
         String type,
