+++ backend/src/main/java/org/donggle/backend/application/service/connection/notion/dto/NotionTokenRequest.java
@@ -1,4 +1,4 @@
-package org.donggle.backend.application.service.oauth.notion.dto;
+package org.donggle.backend.application.service.connection.notion.dto;
 
 public record NotionTokenRequest(String grant_type, String code, String redirect_uri) {
 }
